Unify some common logic of bitbucket   * update snapshot
