+++ source/runner/DangerUtils.ts
@@ -1,6 +1,5 @@
 // The documentation for these are provided inline
 // inside DangerUtilsDSL.ts
-
 export function sentence(array: string[]): string {
   if ((array || []).length === 0) {
     return ""
