+++ source/runner/templates/bitbucketCloudTemplate.ts
@@ -1,21 +1,14 @@
 import { DangerResults } from "../../dsl/DangerResults"
 import { Violation } from "../../dsl/Violation"
+import {
+  dangerIDToString,
+  messageForResultWithIssues,
+  dangerSignature,
+  noEntryEmoji,
+  warningEmoji,
+  messageEmoji,
+} from "./bitbucketTemplateCommon"
 
-// BitBucket Cloud supports these emojis 🎉
-const noEntryEmoji = "❌"
-const warningEmoji = "⚠️"
-const messageEmoji = "✨"
-const signatureEmoji = "🚫"
-
-export const dangerSignature = (results: DangerResults) => {
-  let meta = results.meta || { runtimeName: "dangerJS", runtimeHref: "https://danger.systems/js" }
-  return `Generated by ${signatureEmoji} [${meta.runtimeName}](${meta.runtimeHref})`
-}
-
-export const messageForResultWithIssues = `${warningEmoji}  Danger found some issues. Don't worry, everything is fixable.`
-
-export const dangerIDToString = (id: string) => `danger-id-${id};`
-export const fileLineToString = (file: string, line: number) => `  File: ${file};Line: ${line};`
 /**
  * Postfix signature to be attached comment generated / updated by danger.
  */
@@ -63,20 +56,5 @@ export function template(dangerID: string, results: DangerResults, commitID?: st
   ${dangerSignaturePostfix(results, commitID)}
   
   [](http://${dangerIDToString(dangerID)})
-  `
-}
-
-export function inlineTemplate(dangerID: string, results: DangerResults, file: string, line: number): string {
-  const printViolation = (emoji: string) => (violation: Violation) => {
-    return `- ${emoji} ${violation.message}`
-  }
-
-  return `
-[//]: # (${dangerIDToString(dangerID)})
-[//]: # (${fileLineToString(file, line)})
-${results.fails.map(printViolation(noEntryEmoji)).join("\n")}
-${results.warnings.map(printViolation(warningEmoji)).join("\n")}
-${results.messages.map(printViolation(messageEmoji)).join("\n")}
-${results.markdowns.map(v => v.message).join("\n\n")}
   `
 }
